fix: bug assigning start_ts on restart to oracle
